server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
torch==2.8.0
torchvision==0.21.0
diffusers==0.32.2
transformers==4.48.3
accelerate==1.3.0
safetensors==0.5.2
pillow==11.1.0
pydantic==2.10.6
python-multipart==0.0.20